--- src/routes/onboard/index.svelte
@@ -1,5 +1,0 @@
-<script lang="ts">
-	import Onboard from '../../onboard/Onboard.svelte';
-</script>
-
-<Onboard />
